Allow float transactions amount

+++ common/common/src/main/java/org/shadowrunrussia2020/android/common/models/BalanceResponse.kt
@@ -29,7 +29,7 @@ data class Transaction(
     var operationTime: Date,
     val from: String,
     val to: String,
-    var amount: Int,
+    var amount: Float,
     var comment: String?,
     var transferType: String
 ) : Parcelable
